Move pom related code into a new file: pom.go.
