+++ cli/azd/internal/appdetect/java.go
@@ -2,11 +2,9 @@ package appdetect
 
 import (
 	"context"
-	"encoding/xml"
 	"fmt"
 	"io/fs"
 	"maps"
-	"os"
 	"path/filepath"
 	"slices"
 	"strings"
@@ -59,65 +57,6 @@ func (jd *javaDetector) DetectProject(ctx context.Context, path string, entries 
 	}
 
 	return nil, nil
-}
-
-// mavenProject represents the top-level structure of a Maven POM file.
-type mavenProject struct {
-	XmlName              xml.Name             `xml:"project"`
-	Parent               parent               `xml:"parent"`
-	Modules              []string             `xml:"modules>module"` // Capture the modules
-	Dependencies         []dependency         `xml:"dependencies>dependency"`
-	DependencyManagement dependencyManagement `xml:"dependencyManagement"`
-	Build                build                `xml:"build"`
-	path                 string
-}
-
-// Parent represents the parent POM if this project is a module.
-type parent struct {
-	GroupId    string `xml:"groupId"`
-	ArtifactId string `xml:"artifactId"`
-	Version    string `xml:"version"`
-}
-
-// Dependency represents a single Maven dependency.
-type dependency struct {
-	GroupId    string `xml:"groupId"`
-	ArtifactId string `xml:"artifactId"`
-	Version    string `xml:"version"`
-	Scope      string `xml:"scope,omitempty"`
-}
-
-// DependencyManagement includes a list of dependencies that are managed.
-type dependencyManagement struct {
-	Dependencies []dependency `xml:"dependencies>dependency"`
-}
-
-// Build represents the build configuration which can contain plugins.
-type build struct {
-	Plugins []plugin `xml:"plugins>plugin"`
-}
-
-// Plugin represents a build plugin.
-type plugin struct {
-	GroupId    string `xml:"groupId"`
-	ArtifactId string `xml:"artifactId"`
-	Version    string `xml:"version"`
-}
-
-func readMavenProject(filePath string) (*mavenProject, error) {
-	bytes, err := os.ReadFile(filePath)
-	if err != nil {
-		return nil, err
-	}
-
-	var project mavenProject
-	if err := xml.Unmarshal(bytes, &project); err != nil {
-		return nil, fmt.Errorf("parsing xml: %w", err)
-	}
-
-	project.path = filepath.Dir(filePath)
-
-	return &project, nil
 }
 
 func detectDependencies(mavenProject *mavenProject, project *Project) (*Project, error) {
